requirements.txt
pandas>=1.5.0
PyYAML>=6.0
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
Pillow>=8.0.0
matplotlib>=3.4.0
tqdm>=4.60.0
scikit-learn>=1.0.0
pytorch>=1.10.0
pytorchvision>=0.11.1
numpy>=1.21.0
